Refactor of the scenario class
