+++ protocol/sign/src/androidTest/kotlin/com/walletconnect/sign/test/scenario/SignActivityScenario.kt
@@ -1,0 +1,58 @@
+package com.walletconnect.sign.test.scenario
+
+import androidx.lifecycle.Lifecycle
+import androidx.test.core.app.ActivityScenario
+import com.walletconnect.sign.test.activity.InstrumentedTestActivity
+import junit.framework.TestCase
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
+import timber.log.Timber
+import java.util.concurrent.CountDownLatch
+import java.util.concurrent.TimeUnit
+
+open class SignActivityScenario {
+    var scenario: ActivityScenario<InstrumentedTestActivity>? = null
+    private var scenarioLaunched: Boolean = false
+    private val latch = CountDownLatch(1)
+    private val testScope: CoroutineScope = CoroutineScope(Dispatchers.Default)
+
+    fun initLogging() {
+        if (Timber.treeCount == 0) {
+            Timber.plant(
+                object : Timber.DebugTree() {
+                    override fun log(priority: Int, tag: String?, message: String, t: Throwable?) {
+                        super.log(priority, "WalletConnectV2", message, t)
+                    }
+                }
+            )
+        }
+    }
+
+    fun launch(timeoutSeconds: Long = 1, testCodeBlock: suspend () -> Unit) {
+        require(!scenarioLaunched) { "Scenario has already been launched!" }
+
+        scenario = ActivityScenario.launch(InstrumentedTestActivity::class.java)
+        scenarioLaunched = true
+
+        scenario?.moveToState(Lifecycle.State.RESUMED)
+        assert(scenario?.state?.isAtLeast(Lifecycle.State.RESUMED) == true)
+
+        testScope.launch { testCodeBlock() }
+
+        try {
+            TestCase.assertTrue(latch.await(timeoutSeconds, TimeUnit.SECONDS))
+        } catch (exception: Exception) {
+            TestCase.fail(exception.message)
+        }
+    }
+
+    fun closeAsSuccess() {
+        latch.countDown()
+    }
+
+    fun afterAll() {
+        Timber.d("afterAll")
+        scenario?.close()
+    }
+}
